added country detail screen
