+++ src/index.scss
@@ -2,4 +2,5 @@
 
 body{
     background-color: $light-green;
+    font-family: 'Inter', sans-serif;
 };
