+++ src/screens/mainPage/CountryDetail/styles.module.scss
@@ -1,6 +1,11 @@
-.container {
+@import '/src/scss/variables/colors';
+.container{
     display: flex;
-    justify-content: center;
-    gap: 10px;
-    flex-wrap: wrap;
+    justify-content: space-evenly;
+    
+}
+
+.title{
+    text-align: center;
+    color: $green;
 }
