+++ src/screens/mainPage/RegionDetail/styles.module.scss
@@ -1,0 +1,6 @@
+.container {
+    display: flex;
+    justify-content: center;
+    gap: 10px;
+    flex-wrap: wrap;
+}
